moved model loading code into VRModel class
