+++ spatialize/source/main.cpp
@@ -16,10 +16,6 @@ main(int argc, char** argv)
 {
 	MinVR::AbstractMVREngine *engine = new MinVR::MVREngineGLFW();
 	engine->init(argc, argv);
-    if (!argv[2]) {
-        std::cout << "SPATIALIZE ERROR: No path to model specified." << std::endl;
-        exit(1);
-    }
 	MinVR::AbstractMVRAppRef app(new Spatialize::SpatializeApp(argv[2]));
 	engine->runApp(app);
 	delete engine;
